requirements.txt
numpy
pandas
matplotlib
h5py
scipy
Pillow
pyFAI
